Add img srcs back again from a remote source
Honestly I don't really know what I was doing wrong but I think by using
an absolute link somehow it works on the main page to display the img. And
somehow the img displays correctly on the recipes page. I think it may
have something to do with using the imgs under a list element.